utils/requirements.txt
pillow
numpy
flask
requests
tkinter
huggingface_hub